+++ src/__tests__/dependency-injection.finished.js
@@ -1,0 +1,23 @@
+// these should normally be in your jest setupTestFrameworkScriptFile
+import 'jest-dom/extend-expect'
+import 'react-testing-library/cleanup-after-each'
+
+import React from 'react'
+import {render, fireEvent, wait} from 'react-testing-library'
+import {GreetingLoader} from '../greeting-loader-02-dependency-injection'
+
+test('loads greetings on click', async () => {
+  const mockLoadGreeting = jest.fn(subject =>
+    Promise.resolve({data: {greeting: `Hi ${subject}`}}),
+  )
+  const {getByLabelText, getByText, getByTestId} = render(
+    <GreetingLoader loadGreeting={mockLoadGreeting} />,
+  )
+  const nameInput = getByLabelText(/name/i)
+  const loadButton = getByText(/load/i)
+  nameInput.value = 'Mary'
+  fireEvent.click(loadButton)
+  expect(mockLoadGreeting).toHaveBeenCalledTimes(1)
+  expect(mockLoadGreeting).toHaveBeenCalledWith('Mary')
+  await wait(() => expect(getByTestId('greeting')).toHaveTextContent(`Hi Mary`))
+})
