+++ src/__tests__/mock-component.finished.js
@@ -1,0 +1,26 @@
+// these should normally be in your jest setupTestFrameworkScriptFile
+import 'jest-dom/extend-expect'
+import 'react-testing-library/cleanup-after-each'
+
+import React from 'react'
+import {render, fireEvent} from 'react-testing-library'
+import {HiddenMessage} from '../hidden-message'
+
+jest.mock('react-transition-group', () => {
+  return {
+    CSSTransition: props => (props.in ? props.children : null),
+  }
+})
+
+test('shows hidden message when toggle is clicked', () => {
+  const myMessage = 'hello world'
+  const {getByText, queryByText} = render(
+    <HiddenMessage>{myMessage}</HiddenMessage>,
+  )
+  const toggleButton = getByText(/toggle/i)
+  expect(queryByText(myMessage)).not.toBeInTheDocument()
+  fireEvent.click(toggleButton)
+  expect(getByText(myMessage)).toBeInTheDocument()
+  fireEvent.click(toggleButton)
+  expect(queryByText(myMessage)).not.toBeInTheDocument()
+})
