+++ src/__tests__/unmounting.finished.js
@@ -1,0 +1,24 @@
+// these should normally be in your jest setupTestFrameworkScriptFile
+import 'jest-dom/extend-expect'
+import 'react-testing-library/cleanup-after-each'
+
+import React from 'react'
+import {render} from 'react-testing-library'
+import {Countdown} from '../countdown'
+
+jest.useFakeTimers()
+
+beforeEach(() => {
+  jest.spyOn(console, 'error').mockImplementation(() => {})
+})
+
+afterEach(() => {
+  console.error.mockRestore()
+})
+
+test('does not attempt to set state when unmounted (to prevent memory leaks)', () => {
+  const {unmount} = render(<Countdown />)
+  unmount()
+  jest.runOnlyPendingTimers()
+  expect(console.error).not.toHaveBeenCalled()
+})
